Add FileCommand handling file operations
